requirements.txt
torch>=2.0.0
transformers>=4.40.0
tokenizers>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
joblib>=1.2.0
rdkit-pypi>=2022.9.5
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.2
chembl-webresource-client>=0.10.8
